fast_pinokkio/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
numpy>=1.26.0
opencv-python>=4.8.0
insightface>=0.7.0
aiomysql>=0.2.0
python-dotenv>=0.21.0
aiohttp>=3.8.0
Pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
cryptography>=42.0.8
redis>=5.0.7
onnxruntime
